Move deleteNote from noteModal to noteUtils
